+++ ipware/ip.py
@@ -1,5 +1,5 @@
 from . import utils as util
-from . import defaults as defs
+from django.conf import settings
 
 
 def get_client_ip(
@@ -19,7 +19,7 @@ def get_client_ip(
         proxy_trusted_ips = []
 
     if request_header_order is None:
-        request_header_order = defs.IPWARE_META_PRECEDENCE_ORDER
+        request_header_order = settings.IPWARE_META_PRECEDENCE_ORDER
 
     for key in request_header_order:
         value = util.get_request_meta(request, key)